Change in how appintent functions work Life stats bug fix Documentation improvements

+++ Anki.Vector/Types/RobotLifetimeStats.cs
@@ -32,13 +32,13 @@ namespace Anki.Vector.Types
         public int UtilitiesUsed { get; private set; }
 
         /// <summary>
-        /// Gets the distance driven.
+        /// Gets the distance driven as millions of centimeters.
         /// </summary>
         [JsonProperty("Odom.Body")]
-        public int DistanceDriven { get; private set; }
+        public decimal DistanceDriven { get; private set; }
 
         /// <summary>
-        /// Gets the duration of the petting in seconds.
+        /// Gets the duration of the petting in milliseconds.
         /// </summary>
         [JsonProperty("Pet.ms")]
         public int PettingDuration { get; private set; }
